refactored class base
