+++ mlens/ensemble/stacking_ensemble.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python3
+#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
 """ML-ENSEMBLE
@@ -13,9 +13,9 @@ Scikit-learn API allows full integration, including grid search and pipelining.
 from __future__ import division, print_function
 
 from sklearn.base import clone, BaseEstimator, TransformerMixin, RegressorMixin
-from ._setup import name_estimators, name_base, _split_base
-from ._clone import _clone_base_estimators, _clone_preprocess_cases
-from ._support import _check_estimators, name_columns
+from ..base import name_estimators, name_base, _split_base
+from ..base import _clone_base_estimators, _clone_preprocess_cases
+from ..base import _check_estimators, name_columns
 from ..utils import print_time
 from ..metrics import score_matrix
 from ..parallel import (preprocess_folds, preprocess_pipes,
